old-tests/submission/Microsoft/forms/formtarget.html

<!DOCTYPE html>
<html><head>
        <title>formtarget on button element</title>
        <meta content="text/html; charset=UTF-8" http-equiv="content-type">
        <meta content="formtarget on button element" name="description">
    <link href="http://www.w3.org/TR/html5/attributes-common-to-form-controls.html#dom-fs-formtarget" rel="help">
</head>
    <body>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>

        <div id="log"></div>    
        <button formtarget="_self" style="display : none" type="submit">Submit</button>
        <input formtarget="_self" style="display : none" type="submit" value="submit">
        <input style="display : none" type="submit" value="submit">
        
        <script type="text/javascript">
            test(function() {assert_equals(document.getElementsByTagName("button")[0].formTarget, "_self")}, "formtarget on button element");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].formTarget, "_self")}, "formtarget on input element");
            
            var testElem = document.getElementsByTagName("input")[0];
            testElem.formTarget = "_blank";

            test(function() {assert_equals(document.getElementsByTagName("input")[0].formTarget, "_blank")}, "setting formtarget on input element");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].getAttribute("formtarget"), "_blank")}, "setting formtarget will overwrite attribute value");
            
            test(function() {assert_equals(document.getElementsByTagName("input")[1].formTarget, "")}, "formtarget property default value is an empty string");
            
        </script>
    
</body></html>
